refactor(core): rename package 'entities' to 'domain' and update references
