+++ src/main/java/br/com/carloslonghi/eventflow/api/core/usecases/SearchEventUseCase.java
@@ -1,6 +1,6 @@
 package br.com.carloslonghi.eventflow.api.core.usecases;
 
-import br.com.carloslonghi.eventflow.api.core.entities.Event;
+import br.com.carloslonghi.eventflow.api.core.domain.Event;
 
 public interface SearchEventUseCase {
