+++ src/main/java/br/com/carloslonghi/eventflow/api/infra/mapper/EventDtoMapper.java
@@ -1,13 +1,13 @@
 package br.com.carloslonghi.eventflow.api.infra.mapper;
 
-import br.com.carloslonghi.eventflow.api.core.entities.Event;
+import br.com.carloslonghi.eventflow.api.core.domain.Event;
 import br.com.carloslonghi.eventflow.api.infra.dtos.EventDto;
 import org.mapstruct.Mapper;
 
 @Mapper
 public interface EventDtoMapper {
 
-    EventDto toDto(Event entity);
+    EventDto toDto(Event domain);
 
-    Event toEntity(EventDto dto);
+    Event toDomain(EventDto dto);
 }
