+++ src/main/java/br/com/carloslonghi/eventflow/api/infra/mapper/EventEntityMapper.java
@@ -1,13 +1,13 @@
 package br.com.carloslonghi.eventflow.api.infra.mapper;
 
-import br.com.carloslonghi.eventflow.api.core.entities.Event;
+import br.com.carloslonghi.eventflow.api.core.domain.Event;
 import br.com.carloslonghi.eventflow.api.infra.persistence.EventEntity;
 import org.mapstruct.Mapper;
 
 @Mapper
 public interface EventEntityMapper {
 
-    EventEntity toEntity(Event event);
+    EventEntity toEntity(Event domain);
 
-    Event toDomain(EventEntity eventEntity);
+    Event toDomain(EventEntity entity);
 }
