+++ src/main/java/br/com/carloslonghi/eventflow/api/infra/presentation/EventController.java
@@ -1,7 +1,7 @@
 package br.com.carloslonghi.eventflow.api.infra.presentation;
 
-import br.com.carloslonghi.eventflow.api.core.entities.Event;
-import br.com.carloslonghi.eventflow.api.core.usecases.CreateEventUseCaseImpl;
+import br.com.carloslonghi.eventflow.api.core.domain.Event;
+import br.com.carloslonghi.eventflow.api.core.usecases.CreateEventUseCase;
 import br.com.carloslonghi.eventflow.api.infra.dtos.EventDto;
 import br.com.carloslonghi.eventflow.api.infra.mapper.EventDtoMapper;
 import lombok.RequiredArgsConstructor;
@@ -12,13 +12,12 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class EventController {
 
-    private final CreateEventUseCaseImpl createEventUseCase;
-
+    private final CreateEventUseCase createEventUseCase;
     private final EventDtoMapper eventDtoMapper;
 
     @PostMapping
     public EventDto createEvent(@RequestBody EventDto eventDto) {
-        Event event = eventDtoMapper.toEntity(eventDto);
+        Event event = eventDtoMapper.toDomain(eventDto);
         Event eventCreated = createEventUseCase.execute(event);
 
         return eventDtoMapper.toDto(eventCreated);
